fixed the bounding box of the player

+++ core/src/com/mygdx/graphics/screens/GameScreenInput.java
@@ -16,21 +16,19 @@ public class GameScreenInput implements InputProcessor {
     private OrthographicCamera camera;
     private Player player;
     private TiledMapTileLayer collisionLayer;
-    private float mobileScale = 1;
+    private double mobileScale = 1;
     private boolean isPressedUp, isPressedDown, isPressedRight, isPressedLeft;
 
-    public GameScreenInput(Player player, TiledMapTileLayer collisionLayer)
-    {
+    public GameScreenInput(Player player, TiledMapTileLayer collisionLayer) {
         this.collisionLayer = collisionLayer;
         this.player = player;
         this.camera = player.getCamera();
     }
 
-    public void processInput()
-    {
+    public void processInput() {
         if(isPressedRight)
             player.setVelocityX(player.getMovementSpeed());
-        if(isPressedLeft )
+        if(isPressedLeft)
             player.setVelocityX(-player.getMovementSpeed());
         if(isPressedUp)
             player.setVelocityY(player.getMovementSpeed());
@@ -52,24 +50,21 @@ public class GameScreenInput implements InputProcessor {
     }
 
     @Override
-    public boolean keyUp(int keycode) {
-        if(keycode == Input.Keys.A|| keycode == Input.Keys.LEFT)
-        {
+    public boolean keyUp(int keycode)
+    {
+        if(keycode == Input.Keys.A|| keycode == Input.Keys.LEFT) {
             player.setVelocityX(0);
             isPressedLeft = false;
         }
-        if(keycode == Input.Keys.D || keycode == Input.Keys.RIGHT)
-        {
+        if(keycode == Input.Keys.D || keycode == Input.Keys.RIGHT) {
             player.setVelocityX(0);
             isPressedRight = false;
         }
-        if(keycode == Input.Keys.W || keycode == Input.Keys.UP)
-        {
+        if(keycode == Input.Keys.W || keycode == Input.Keys.UP) {
             player.setVelocityY(0);
             isPressedUp = false;
         }
-        if(keycode == Input.Keys.S || keycode == Input.Keys.DOWN)
-        {
+        if(keycode == Input.Keys.S || keycode == Input.Keys.DOWN) {
             player.setVelocityY(0);
             isPressedDown = false;
         }
